+++ CHANGELOG.md
@@ -20,6 +20,8 @@
     
 * Features:
   * Added Timeout for Page::getFullPageClip (thanks @ArseniyShestakov)
+  * Added options `headerTemplate` and `footerTempalte` for Page::pdf (thanks @MUGagency)
+  * Added options `headerTemplate` and `footerTempalte` for Page::pdf (thanks @MUGagency)
 * Bug fixes:
   * none